requirements.txt
alembic==1.3.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
deprecation==2.0.7
idna==2.8
Mako==1.1.0
MarkupSafe==1.1.1
packaging==19.2
PyOTRS==0.4.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.1.2
psycopg2==2.7.5
requests==2.22.0
schedule==0.6.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
